+++ test/protractor/boards-create.js
@@ -15,6 +15,9 @@ describe('Boards overview screen', function () {
     expect(element(by.css('.sidebar-menu a[href="/projects"]')).isPresent()).toBe(true);
     element(by.css('.sidebar-menu a[href="/projects"]')).click();
 
+    // just in case if previous tests didn't clear the search
+    element(by.model('search.name')).clear();
+
     expect(element(by.css('.table.table-hover a[href="/projects/2"]')).isPresent()).toBe(true);
     element(by.css('.table.table-hover a[href="/projects/2"]')).click();